+++ src/chat_outreach_engine/adapters/__init__.py
@@ -4,12 +4,14 @@ from .gorgias import GorgiasAdapter
 from .helpscout import HelpScoutAdapter
 from .intercom import IntercomAdapter
 from .livechat import LiveChatAdapter
+from .olark import OlarkAdapter
 from .reamaze import ReamazeAdapter
 from .shopify_inbox import ShopifyInboxAdapter
 from .tawk import TawkAdapter
 from .tidio import TidioAdapter
 from .zendesk import ZendeskAdapter
+from .zoho_salesiq import ZohoSalesIQAdapter
 
 __all__ = ["ChatraAdapter", "CrispAdapter", "GorgiasAdapter", "HelpScoutAdapter", "IntercomAdapter",
-           "LiveChatAdapter", "ReamazeAdapter", "ShopifyInboxAdapter", "TawkAdapter", "TidioAdapter",
-           "ZendeskAdapter"]
+           "LiveChatAdapter", "OlarkAdapter", "ReamazeAdapter", "ShopifyInboxAdapter", "TawkAdapter",
+           "TidioAdapter", "ZendeskAdapter", "ZohoSalesIQAdapter"]